picking

+++ engine/modules/ECS/include/ecs/entity.hpp
@@ -5,7 +5,7 @@
 
 namespace astre::ecs
 {
-    using Entity = std::uint32_t;
+    using Entity = std::uint64_t;
 
     constexpr Entity INVALID_ENTITY = 0;
     constexpr std::size_t MAX_COMPONENT_TYPES = 256;
